refactor: seperate our trend analysis and visualization into utils, analysis , and visualization modules for better reuse and integration with dashboard
