+++ src/analysis/__init__.py
@@ -1,0 +1,11 @@
+from .trend_analysis import (
+    analyze_award_types,
+    analyze_funding_trends,
+    analyze_sector_funding,
+)
+
+__all__ = [
+    "analyze_funding_trends",
+    "analyze_sector_funding",
+    "analyze_award_types",
+]
