+++ src/visualization/__init__.py
@@ -1,0 +1,11 @@
+from .trend_plotting import (
+    plot_award_types_subplots,
+    plot_funding_trends_subplots,
+    plot_sector_funding_subplots,
+)
+
+__all__ = [
+    "plot_award_types_subplots",
+    "plot_sector_funding_subplots",
+    "plot_funding_trends_subplots",
+]
